+++ src/inputParser.h
@@ -2,6 +2,7 @@
 #define INPUTPARSER_H
 
 #include <string>
+#include "Definitions.h"
 
 #define FREE_SLIP	1
 #define NO_SLIP		2
@@ -11,29 +12,29 @@
 struct ProblemParameters
 {
 	// geometry data
-	double		xlength,		//! domain size in x-direction
+	REAL		xlength,		//! domain size in x-direction
 				ylength;		//! domain size in y-direction
 
 	int			nx,				//! number of interior cells in x-direction
 				ny;				//! number of interior cells in y-direction
 
 	// time stepping data
-	double		dt,				//! time step size
+	REAL		dt,				//! time step size
 				tau;			//! safety factor for time step size control
 
 	// pressure-iteration data
 	int			it_max;			//! maximal number of pressure iterations per time step
 
-	double		epsilon,		//! stopping tolerance eps for pressure iteration
+	REAL		epsilon,		//! stopping tolerance eps for pressure iteration
 				omega,			//! relaxation parameter for SOR iteration
 				gamma;			//! upwind differencing factor
 
 	// problem dependent quantities
-	double		re,				//! Reynolds number Re
+	REAL		re,				//! Reynolds number Re
 				gx,				//! body force gx (e.g. gravity)
 				gy;				//! body force gy (e.g. gravity)
 
-	double		ui,				//! initial velocity in x-direction
+	REAL		ui,				//! initial velocity in x-direction
 				vi,				//! initial velocity in y-direction
 				pi;				//! initial pressure
